+++ apps/web/src/app/nonce/layout.tsx
@@ -3,6 +3,7 @@ import { headers } from "next/headers";
 
 export default function Layout({ children }: { children: React.ReactNode }) {
     const nonce = headers().get("x-nonce");
+    console.log("nonce", nonce);
 
     return <ThemeProvider nonce={nonce}>{children}</ThemeProvider>;
 }