+++ apps/web/src/middleware.ts
@@ -32,5 +32,5 @@ export function middleware(request: NextRequest) {
 }
 
 export const config = {
-    matcher: "/**/nonce"
+    matcher: ["/nonce", "/build/nonce"]
 };